adds page folder with home page, landing page and login page. Updates tests and snapshots. Adds the plant images to the repo
